Move ResultCodes.java to base package
