+++ auth/src/main/java/com/firebase/ui/auth/ui/email/EmailHintContainerActivity.java
@@ -21,10 +21,10 @@ import android.content.IntentSender;
 import android.os.Bundle;
 import android.util.Log;
 
+import com.firebase.ui.auth.ResultCodes;
 import com.firebase.ui.auth.ui.AppCompatBase;
 import com.firebase.ui.auth.ui.BaseHelper;
 import com.firebase.ui.auth.ui.FlowParameters;
-import com.firebase.ui.auth.ui.ResultCodes;
 import com.firebase.ui.auth.util.FirebaseAuthWrapper;
 import com.firebase.ui.auth.util.FirebaseAuthWrapperFactory;
 import com.google.android.gms.auth.api.credentials.Credential;
